+++ test/components/pages/home.test.js
@@ -1,20 +1,11 @@
 import React from 'react';
-import MemoryRouter from 'react-router-dom/MemoryRouter';
 
-import { shallow } from 'enzyme';
-import { ApolloProvider } from 'react-apollo';
-import client from '@/api/apollo';
+import { mount } from '../../test-utils/render';
 import Home from '@/components/pages/home';
 
 describe('Components: <Home />', () => {
   test('renders without exploding', () => {
-    const wrapper = shallow(
-      <MemoryRouter>
-        <ApolloProvider client={client}>
-          <Home />
-        </ApolloProvider>
-      </MemoryRouter>,
-    );
+    const wrapper = mount(<Home />);
     expect(wrapper.html()).toContain('div');
   });
 });
